subprocess kernel manager
